requirements.txt

anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
faiss-cpu==1.12.0
fastapi==0.95.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.3.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.3
gunicorn==20.1.0
h11==0.16.0
idna==3.10
Jinja2==3.1.2
MarkupSafe==3.0.2
numpy==2.3.2
packaging==25.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==1.10.12
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.26.1
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.20.0
